refactor: 残り全モジュールのテスト分離(18ファイル)
Agent使用により以下を一括処理:
- infra(api,storage)層の6モジュールのテスト分離
- domain層の12モジュールのテスト分離
- src/conftest.py削除
- pyproject.tomlをtests/のみに変更
- 約1,500行のTestModクラスを削除

修正内容:
- import文の正規化(src.プレフィックス削除)
- インデント修正(test_command.py, test_search.py)
- 型ヒント追加(tmp_path: Path)
- conftest.pyにテーブル定義インポート追加

テスト結果: 53 passed, 11 failed (残りエラーは別途修正予定)
