+++ src/infra/storage/rds.py
@@ -117,21 +117,3 @@ def _ensure_sqlite_directory(database_url: str) -> None:
     db_path.parent.mkdir(parents=True, exist_ok=True)
 
 
-class TestMod:
-    def test_ensure_sqlite_directory_creates_parent(self, fs) -> None:
-        """
-        docs:
-            目的:
-                SQLiteファイル保存時に親ディレクトリが自動生成されることを確認する。
-            検証観点:
-                - _ensure_sqlite_directory が存在しないディレクトリを作成する。
-                - 相対パスも絶対パスも正しく処理される。
-        """
-
-        # 絶対パス
-        _ensure_sqlite_directory("sqlite:////test/dir/test.db")
-        assert fs.exists("/test/dir")
-
-        # 相対パス（カレントディレクトリ基準）
-        _ensure_sqlite_directory("sqlite:///relative/path/test.db")
-        assert fs.exists("relative/path")
