+++ tests/domain/task_queue/http_request/test_command.py
@@ -1,0 +1,51 @@
+"""domain.task_queue.http_request.command のテスト"""
+
+from datetime import datetime
+
+from sqlmodel import select
+
+from domain.common import ensure_saved_at
+from domain.task_queue.http_request.command import store_http_request
+from domain.task_queue.http_request.model import HttpRequestTaskRecord
+from domain.task_queue.http_request.service import create_http_request
+from infra.storage.rds import session_scope
+
+
+def test_store_http_request_persists_record() -> None:
+    """
+    docs:
+    目的:
+        store_http_request が永続化を行い、戻り値として最新状態の
+        HttpRequestTask を返すことを確認する。
+    検証観点:
+        - create_http_request で生成した HttpRequestTask が
+          store_http_request で保存される。
+        - 保存後に同一IDのレコードがDB上に存在する。
+        - created_at が保持され、updated_at が更新される。
+    """
+
+    # pytestにより自動的にインメモリDBが使用される
+    request = create_http_request(
+        url="https://example.com/store",
+        description="Store Request",
+        group="test:store",
+        status_code=201,
+        created_at=datetime(2024, 1, 1, 9, 0, 0),
+    )
+
+    stored = store_http_request(request)
+    assert stored.id == request.id
+    assert stored.status_code == request.status_code
+    assert stored.created_at == ensure_saved_at(datetime(2024, 1, 1, 9, 0, 0))
+    assert stored.updated_at >= stored.created_at
+
+    with session_scope() as session:
+        statement = select(HttpRequestTaskRecord).where(
+            HttpRequestTaskRecord.id == request.id
+        )
+        record = session.exec(statement).first()
+        assert record is not None
+        assert record.description == "Store Request"
+        assert record.group == "test:store"
+        assert ensure_saved_at(record.created_at) == stored.created_at
+        assert ensure_saved_at(record.updated_at) == stored.updated_at
